build UserSearchBar component functionality
