+++ src/service/apiClient.js
@@ -29,7 +29,7 @@ async function postPost(newPost) {
 }
 
 async function getUserByName(firstName) {
-  const res = await get(`users?first_name=${firstName}`);
+  const res = await get(`users?name=${firstName}`);
   return res.data.users;
 }
